starting to remove functions into utils
