+++ lib/plugins/css_plugin.js
@@ -1,4 +1,4 @@
-import glob from 'glob'
+import util from './plugin_utils'
 import path from 'path'
 import SingleEntryDependency from 'webpack/lib/dependencies/SingleEntryDependency'
 
@@ -14,8 +14,7 @@ export default class CSSWebpackPlugin {
 
   apply (compiler) {
     // read file tree and get all css files
-    this.files = glob.sync(`${compiler.options.context}/${this.opts.matcher}`)
-                    .filter(this._removeIgnores.bind(this))
+    this.files = util.getFilesFromGlob(compiler, this.opts)
 
     // inject css files into webpack's pipeline
     compiler.plugin('make', (compilation, done) => {
