+++ lib/plugins/jade_plugin.js
@@ -1,35 +1,21 @@
-import glob from 'glob'
+import util from './plugin_utils'
 import path from 'path'
-import SingleEntryDependency from 'webpack/lib/dependencies/SingleEntryDependency'
 
 export default class JadeWebpackPlugin {
 
   constructor (opts) {
+    // TODO use joi for this
     if (!opts) { this.opts = {} } else { this.opts = opts };
   }
 
   apply (compiler) {
     // read file tree and get all jade files
-    this.files = glob.sync(`${compiler.options.context}${this.opts.matcher}`)
-                    .filter(this._removeIgnores.bind(this))
+    this.files = util.getFilesFromGlob(compiler, this.opts)
 
     // inject jade files into webpack's pipeline
     compiler.plugin('make', (compilation, done) => {
-      let tasks = []
-
-      this.files.forEach(f => {
-        let name = f.match(/\/(\w+).jade/)[1]
-        let relativePath = f.replace(compiler.options.context, '.')
-        let dep = new SingleEntryDependency(relativePath)
-
-        tasks.push(new Promise((resolve, reject) => {
-          compilation.addEntry(compiler.options.context, dep, name, (err) => {
-            if (err) { reject(err) } else { resolve(true) }
-          })
-        }))
-      })
-
-      Promise.all(tasks).then((res) => { done() })
+      util.addFilesAsWebpackEntries(this.files, this.opts, compiler, compilation)
+          .then(res => done())
     })
 
     // grab the sources and dependencies and export them into the right files
@@ -66,13 +52,6 @@ export default class JadeWebpackPlugin {
     })
 
     return res.substring(1).replace(/\.jade$/, '.html')
-  }
-
-  _removeIgnores (f) {
-    for (let ignore of this.opts.ignore) {
-      if (f.match(ignore)) { return false }
-    }
-    return true
   }
 
 }
